+++ frontend/src/App.tsx
@@ -1,7 +1,5 @@
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
 
-import useToggle from '@/hooks/useToggle';
-
 import NavBar from '@/components/NavBar/NavBar';
 import ProtectRoute from '@/components/ProtectRoute';
 import SideBar from '@/components/SideBar/SideBar';
@@ -12,21 +10,17 @@ import MyPage from '@/pages/MyPage/MyPage';
 
 import { PATH } from '@/constants';
 
-function App() {
-  const { state: isLoginModalOpen, toggleState: toggleLoginModalOpen } = useToggle();
+import LoginPage from './pages/LoginPage/LoginPage';
 
+function App() {
   return (
     <Router>
-      <NavBar openLoginModal={toggleLoginModalOpen} />
+      <NavBar />
       <SideBar />
       <Routes>
-        <Route
-          path={PATH.MAIN}
-          element={
-            <MainPage isLoginModalOpen={isLoginModalOpen} closeLoginModal={toggleLoginModalOpen} />
-          }
-        />
+        <Route path={PATH.MAIN} element={<MainPage />} />
         <Route path={PATH.AUTH} element={<AuthPage />} />
+        <Route path={PATH.LOGIN} element={<LoginPage />} />
         <Route element={<ProtectRoute />}>
           <Route path={PATH.CATEGORY} element={<CategoryPage />} />
           <Route path={PATH.PROFILE} element={<MyPage />} />
