+++ frontend/src/components/CategoryAddModal/CategoryAddModal.styles.ts
@@ -31,6 +31,7 @@ const content = ({ flex }: Theme) => css`
 
   width: 100%;
   height: 100%;
+
   justify-content: center;
 `;
