+++ frontend/src/components/NavBar/NavBar.stories.tsx
@@ -7,6 +7,6 @@ export default {
   component: NavBar,
 } as ComponentMeta<typeof NavBar>;
 
-const Template: ComponentStory<typeof NavBar> = (args) => <NavBar {...args} />;
+const Template: ComponentStory<typeof NavBar> = () => <NavBar />;
 
 export const Primary = Template.bind({});
